MMFMWK-7894-4:select() has crash issue on android
When calling select() or pselect() in dsp wrapper on
android platform, "blocked syscall: pselect6" error
will occur.

This issue is caused by parameter compatibility when
transferring data between user space and kernel. Because
dsp wrapper is compiled as 32-bit while kernel is compiled
64-bit. The length of fd_set type which is used by select()
is different between 32-bit toolchain and 64-bit toolchain,
which leads to this issue.

While there is no this parameter compatibility problem
for poll() function, so use poll() instead of select().

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>