MMFMWK-8061: proxy: Don't use a timeout waiting for responses
Because xf_proxy_thread is receiving messages asynchronously it shouldn't
time out..

One reason for this is that for example, there might be no response to
wait for.

Have a look at the following example:

int main(void)
{
	xaf_adev_open(dev);
	sleep(5);

	/* .. */
	return 0;
}

This will terminate xf_proxy_thread because no message is sent to DSP.

A more important use case where xf_proxy_thread is terminated is
when using gplay:
	* press 'a' wait 5 seconds and then press 'a' again to go back
	to play
        * or for example if gplay finishes decoding very quickly and
	does some other action. Then it would be able to do a proper
	cleanup.

The only drawback for this patch, is what happens if there is a bug in
DSP and a message is really lost? In this case, I think we should see
some warnings in DSP driver and investigate there.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>