MMFMWK-8522-0 dsp_wrapper/: Fix issue no vioce when play aac streams.
When play aac streams in Andriod, it hang at the beginning for short
time. if seek to nearly end, and play to wait for eos event, the audio
decode did not stop and wait for long time to stop decoding. And only
can hear noise.
The reason that cause issue is that not set parameter FRAMED, dsp
wrapper will not pass this parameter to dsp before, now changed it and
dsp aac ua_wrapper can receive the parameter.

Signed-off-by: Zhang Peng <peng_zhang_8@nxp.com>