MMFMWK-8983-1: common: proxy: refine the flow for closing remoteproc
As we support two instance, the close operation is
a problem, need to make sure no instance is running,
then we can stop remoteproc, this protection should
be done in kernel, but there is issue for kernel
common code, so we temporaly workaround it in user
space. later if the issue in kernel fix, this commit
can be reverted.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>