LF-6852 dsp_wrapper: Fix memory leak issue
Find the memory can't stop to increase if looping a playback with dsp.
The reason is not free the output buffer in wrapper when no output.

Now we handle output buffer as such: if malloc output buffer in
wrapper, there are no output then free in wrapper or should
free in top layer.

Signed-off-by: Zhang Peng <peng.zhang_8@nxp.com>